CommonJS only
Supporting <script> and RequireJS forms simultaneously from the same
source as a CommonJS package complicates dependency management.  Reusing
external dependencies, particularly the LruMap in the collections
package, instead of embedding an implementation, crosses a threshhold,
beyond which manual dependency management is not practical.

This commit switches to using the collections package and bumps the
version so anyone depending on 0.3.x with a RequireJS repository may
keep the old version.  I do not expect many projects to be impacted by
this change.