Add a sanitizeError option and specs
Error objects often contain information, such as the stack, that
should not be exposed to the client.

This adds the ability to give a `sanitizeError` function to the
Connection that will be called whenever an Error is encoded. The
return value of this function is used in place of the original error.